Use the default-config.yml as default configuration
The EmptyConfig with the old behaviour of isActive check being true on default could lead to scared users. No one wants to use a tool which reports that your project has hundreds of code smells. The motivation bar is low. By just using the default configuration is a subset of all rules, we can achieve better adoption.