Error message agrees with require condition
I just got bit by this. I had a "detekt.yaml" configuration and the error told me it either did not exist or it did not end on ".yaml"

The condition check actually checked that the file ends on ".yml" - no "a", but the error was misleading me to think that it couldn't find it because the ending was correct.