More idiomatic build configuration
- All repositories required for building moved to the `settings.gradle`
- Only apply plugins when they are used
- Remove deplicate plugin requests by removing some `buildscript` blocks
- Compile all `KotlinCompile` tasks together with same configuration

Some of these changes will allow for easier migration to the `kotlin-dsl`.
In the `kotlin-dsl`, you want all plugins to be declared in the `plugins {}` blocks so that accessors can be generated.

The `detekt-sample-extensions` project should be considered as another example project or possibly another build to run with composite builds.