Rule `Explicit it lambda parameter` (#1197)
* Add new ExplicitItParameter rule

* Add ExplicitItLambdaParameter rule to styles RuleSet

* Add documentation for ExplicitItLambdaParameter rule

* Update Mishkun on contrib list

* Add common keywords file to be used in rules package

Many keyword literals are duplicated in rules, there is no source of truth.
To hadle that, make the file with common keywords file and extracted keywords from my authored rules there for example

* Add static import for Assertions.assertThat

Improves style and code aesthetic

* Make tests more sound by better separating test cases in groups

Rearrange tests by tree structure:
├ one parameter lambdas
│ ├ explicit `it` parameter
│ │ ├ with parameter type
│ │ └ without parameter type
│ └ implicit `it` parameter
└ multiple parameter lambdas
  ├ with parameter type
  └ without parameter type

* Report ExplicitIt even if it is one of multiple parameters of lambda

Problem:
ExplicitItLambdaParameter rule report ony usage of explicit `it` as a
single parameter in lambda. But more dangerous and confusing is using
`it` as one of the parameters of the multiparameter lambda. This usecase
is not flagged yet

Solution:
 - Make the rule flag any usage of `it` as a lambda parameter.
 - Update documentation to include example of multiparameter lambda
 - Inverse tests for multiparameter lambda

* Fix typo in ExplicitItLambdaParameter rule docs

* Regenerate docs for ExplicitItLambdaParameter rule

* Make rule docs more formal