Publish detekt-api docs to the detekt documentation site (#1615)
* Don't link to docs that don't exist

* Add generateDocumentation task to documentation task group

* Only process KDoc for detekt-api

* Generate empty javadoc artifacts

* Link documentation to JDK 8

* Publish detekt-api docs to docs site

* Use relative URLs when linking internally in docs

Since KDoc output is in a subfolder in the pages directory, previous
assumptions about how to access static resources no longer hold. Instead
it's necessary to use relative URLs which will resolve to the correct
URL regardless of whether the site is hosted locally or on GitHub pages.

* Don't remove slashes within generated URL

Pages in subfolders need the slashes within the URL to properly specify the
folder path.