Create a task per source set (#1597)
* Print stacktrace when build fails with exception

* Add a module name for the Kotlin environment

* Create BindingContext based on environment source files

The compiler configuration now holds a reference to all source files
that will be analyzed in addition to the classpaths required to complete
analysis. This required some changes to when and how the BindingContext
was generated.

Now, the KtFiles that are going to be analyzed are created by the Kotlin
compiler itself and don't need to be compiled by detekt. To ensure
autoCorrect options are respected the files will still be compiled by
detekt when autoCorrect is enabled.

* Create a task per JVM source set & pass required CLI flags

* Add JVM target CLI option for type & symbol resolution

* Only show compiler messages if they're important

* fixes