Use kotlin plugin in gradle integration tests.
Since the kotlin plugin is very likely to be present in projects using
detekt and since we use it to add tasks for source sets we should
use it in all tests as well.
Add stdlib to test project dependencies as the check task automatically
builds the sources if kotlin plugin is present.