Do not assert number of classes messages on the output anymore
As we have changed detekt to be silent by default, the number of classes in not printed anymore.
We are testing the multi module integration and assert if detekt creates all different kind of reports. Testing the number of classes should not be the primary goal of this test cases.