Refactor FindingsReport + FileBasedFindingsReport
This change removes complexity from both report classes.

If there are no findings reported, both classes return null.
This is done before the printing logic is executed.
Thus, we don't need if-else conditionals for this case.

The debt of existing issues can't be null.
This is guaranteed by the Debt class.