Correct the generated test Gradle build files so they're valid
The previous commit uncovers a problem with the Gradle build files that are
being generated for tests, since a default Kotlin version is no longer
declared on the classpath, and the Kotlin plugin cannot be applied without
a specific version.

Since the Kotlin plugin is not currently used in any tests, and hard coding
a Kotlin version adds an additional maintenance burden, it's better to
remove it for now.

The java-library plugin is added instead so the "implementation" and "api"
configurations will still be available for dependency declaration in tests.