Switch order of correctable rules and other rules
Interestingly, the previous code caused a crash in
(#2693). The problem is that we run Ktlint/CommentSpacing
before running LongMethod. Because CommentSpacing
mutates the ktfile.text, but does not update
ktfile.viewerProvider.document, resulting in
inconsistency and crash.

Note that this commit is merely a short-term fix.
In the long term, we should think be more careful in
rule running order, while leveraging parallelism
to optimize the runtime.