Suppress RedundantVisibilityModifierRule if explicit API mode enabled
This solution uses the languageVersionSettings of the compiler to fetch the API mode setting.

Explicit API mode was added in Kotlin 1.4
It prevents libraries' authors from making APIs public unintentionally.
In this mode, the visibility modifier should be defined explicitly even if it is public.
See: https://kotlinlang.org/docs/reference/whatsnew14.html#explicit-api-mode-for-library-authors

Closes #3125