Allow using regular expressions when defining license header templates (#3486)
Problem: different projects might impose different rules on license headers.
It's not uncommon, for example, that a year in the license header is the year when the source
file was created. This leads to a situation when different source files have different license
headers (at least some parts of them differ) making it impossible to use a single static
license header template to match against.

This change makes it possible to define license header templates as regexps.
An example of such template is:
  \/\/ Copyright 20[0-9]{2} Artur Bosch & Contributors
which would match both
  // Copyright 2020 Artur Bosch & Contributors
and
  // Copyright 2021 Artur Bosch & Contributors

The way `AbsentOrWrongFileLicense` works now is controlled
by the new configuration option `licenseTemplateIsRegex`:
when `licenseTemplateIsRegex` is false the template is considered to be static
and the old matching rules apply.
If `licenseTemplateIsRegex` is set to true the new regexp based logic kicks in.

To avoid breakage of existing projects, the added configuration option
`licenseTemplateIsRegex` is disabled by default.

This commit implements a feature requested here #3100.