Fix test snippet compilation issues
Though isolated test execution of the detekt-rules-complexity module
worked in previous version of LongParameterList improvement, another
step of the building process less lenient than the previous one failed
because the annotations that were used were not properly configured to
target value parameters. A new annotation had to be added to the test
configuration to work around that issue.

Additionally, another test case scenario was added to improve test
coverage.