Include custom get/set operators to ExplicitCollectionElementAccessMethod rule
This actually simplifies code, since any get/set operator can be treated the same; no need to look up type names. `Map.put()` is a type-specific edge case, which can be replaced by `set`, and thus indexed access, when the return value is not needed.

Right now, no check is done whether the return value is use. For both `set` and `put` no code smell should be reported if this is the case.