Don't specify Gradle Enterprise Gradle Plugin version
Checking if this is up to date is not as simple as it should be as the
dependencyUpdates task doesn't identify when it's out of date, so manual
checks have to be done.

Removing the specific version means when the Gradle version is updated,
a recent Gradle Enterprise Gradle Plugin version will come with it,
without having to update it separately.