Enable NestedScopeFunctions with custom functions
This PR demonstrates that when a custom value is specified for the functions
configuration property, there are many false positives from totally unrelated
function calls. This doesn't seem to happen when `functions` is unset in the
configuration, and also seems to be somewhat unreliable in the false positives
it generates.