Document and test edge cases for ForbiddenMethodCall function signatures (#5495)
I was stuck trying to match some function signatures (from Timber on Android)
for the ForbiddenMethodCall rule and had to test inside detekt to figure out how
to do it properly. To make this more clear in the future, I've expanded the
description of the methods configuration parameter and added more tests for
these cases. Some of this might be obvious to some but it was not to me at the
time so I think having some quick tips might be helpful; or perhaps a link to a
more comprehensive guide on Kotlin function signatures (if one exists).