Fix KDocReferencesNonPublicProperty false positive (#5534)
KDocReferencesNonPublicProperty has a common false-positive where there is a name overlap with a public and non-public property (e.g. a local variable). This should not be reported as an issue since the kdoc is referencing the public field.

To fix this, we need to add awareness of the public properties and filter out name overlaps when checking if non-public properties are being referenced in the kdoc. This also includes properties declared as constructor parameters.

Also add a test that constructor parameters are allowed to be referenced in the class kdoc even when they are private fields.