Replaced "threshold" property of LongMethod rule by "allowedLines" (#3679)
The rule reported an issue when the lines of a method is greater or equal the value defined for the  "threshold" property.
It is intended that the value defined in the rule is the maximum allowed lines. So the property is renamed to
"allowedLines" and the check for reporting an issue is changed to only check for greater, no longer for equal.