Replaced "threshold" property of CognitiveComplexMethod rule by "allowedComplexity" (#3679)
The rule reported an issue when the complexity of a method is greater or equal the value defined for the  "threshold" property.
It is intended that the value defined in the rule is the maximum allowed complexity. So the property is renamed to
"allowedComplexity" and the check for reporting an issue is changed to only check for greater, no longer for equal.