Replaced "threshold" property of ComplexCondition rule by "allowedComplexity" (#3679)
The rule reported an issue when the complexity of conditions is greater or equal the value defined for the "threshold" property.
It is intended that the value defined in the rule is the maximum allowed complexity. So the property is renamed to
"allowedComplexity" and the check for reporting an issue is changed to only use greater and no longer equal.