Replaced "functionThreshold" and "constructorThreshold" properties of LongParameterList rule by "allowedFunctionParameters" and "allowedConstructorParameters" (#3679)
The rule reported an issue when the amount of parameters of a method or constructor is greater or equal the value defined for the  "XThreshold" properties.
It is intended that the value defined in the rule is the maximum allowed parameter amount. So the properties are renamed to
"allowedXParameters" and the check for reporting an issue is changed to only check for greater, no longer for equal.