Replaced "threshold" property of MethodOverloading rule by "allowedOverloads" (#3679)
The rule reported an issue when the amount of overloads of methods is greater or equal the value defined for the "threshold" property.
It is intended that the value defined in the rule is the maximum allowed overloads. So the property is renamed to
"allowedOverloads" and the check for reporting an issue is changed to only use greater and no longer equal.