Replaced "threshold" property of NestedBlockDepth rule by "allowedDepth" (#3679)
The rule reported an issue when the nested depth of a function is greater or equal the value defined for the "threshold" property.
It is intended that the value defined in the rule is the maximum allowed nested depth. So the property is renamed to
"allowedDepth" and the check for reporting an issue is changed to only use greater and no longer equal.