Replaced "threshold" property of StringLiteralDuplication rule by "allowedDuplications" (#3679)
The rule reported an issue when the amount of duplications is greater or equal the value defined for the "threshold" property.
It is intended that the value defined in the rule is the maximum allowed duplications. So the property is renamed to
"allowedDuplications" and the check for reporting an issue is changed to only use greater and no longer equal.