Replaced "thresholdX" property of TooManyFunctionsrule by "allowedFunctionsX" (#3679)
The rule reported an issue when the number of functions is greater or equal the value defined for the "threshold" property.
It is intended that the value defined in the rule is the maximum allowed number of functions. So the property is renamed to
"allowedFunctionsX" and the check for reporting an issue is changed to only use greater and no longer equal.