Replaced "threshold" property of CouldBeSequence rule by "allowedOperations" (#3679)
The rule reported an issue when the numbe of operations on a collection is greater or equal the value defined for the "threshold" property.
It is intended that the value defined in the rule is the maximum allowed number of operations. So the property is renamed to
"allowedOperations" and the check for reporting an issue is changed to only use greater and no longer equal.