Replaced "threshold" property of LargeClass rule by "allowedLines" (#3679) (#6188)
The rule reported an issue when the amount of lines of a class  is greater or equal the value defined for the "threshold" property.
It is intended that the value defined in the rule are the maximum allowed lines. So the property is renamed to
"allowedLines" and the check for reporting an issue is changed to only use greater and no longer equal.

Co-authored-by: Matthias Schenk <matthias.schenk@sevdesk.de>