Replaced "threshold" property of CyclomaticComplexMethod rule by "allowedComplexity" (#6189)
* Replaced "threshold" property of CyclomaticComplexMethod rule by "allowedComplexity" (#3679)

The rule reported an issue when the cyclomatic complexity of methods is greater or equal the value defined for the "threshold" property.
It is intended that the value defined in the rule is the maximum allowed complexity. So the property is renamed to
"allowedComplexity" and the check for reporting an issue is changed to only use greater and no longer equal.

* Changed the default value for allowedComplexity in order to not allow greater complexity.

* Correction according to the code review

---------

Co-authored-by: Matthias Schenk <matthias.schenk@sevdesk.de>