Replaced "threshold" property of ComplexInterface rule by "allowedDefinitions" (#3679) (#6185)
The rule reported an issue when the amount of definitions of a method is greater or equal the value defined for the  "threshold" property.
It is intended that the value defined in the rule is the maximum allowed definitions. So the property is renamed to
"allowedDefinitions" and the check for reporting an issue is changed to only check for greater and no longer for equal.

Co-authored-by: Matthias Schenk <matthias.schenk@sevdesk.de>