Return path instead of name in absolutePath()
This previously worked because a full path was used as the filename, when
it should have been the file name alone. Now the PsiFile is sourced from
the compiler environment the `name` field contains only the file name, so
this is updated to reference the virtual file's path on the file system.