Remove test for baseline = null
It's not possible to set the value of a RegularFileProperty to null due to
overload resolution ambiguity errors from Kotlin compiler.

Removing the ability to set to null should be an acceptable trade off to
allow the usage of the more strongly typed RegularFileProperty. It is also
likely that this property will be removed from the extension in favour of
just using the task level properties.