Fix issues related to kotlin-stdlib's Gradle module metadata in Kotlin 1.9.20 (#6590)
* Remove redundant dependency

Test fixtures aren't published and there's no need to specify this
dependency for internal consumers.

* Use latest annotations version in compile classpath

kotlin-stdlib is now published with Gradle module metadata which has
caused a new issue because we use consistent dependency resolution (see
useCompileClasspathVersions). Updating the version of the dependency in the
compile classpath fixes it.

KT-49803 is already open to request removing annotations as a transitive
dependency. Until that's fixed this setup will allow builds to operate
correctly.