Use consistent JDK for Kotlin compilation
If toolchains are not used the version of Java that Gradle is running with
is passed as an input to Kotlin compilation tasks. Using toolchains means
compilation tasks have a consistent JDK version as the input. Builds can
then run on any Java version and hit remote cache primed with builds run on
different Java versions.