Use consistent JDK for Kotlin compilation (#6706)
* Use consistent JDK for Kotlin compilation

If toolchains are not used the version of Java that Gradle is running with
is passed as an input to Kotlin compilation tasks. Using toolchains means
compilation tasks have a consistent JDK version as the input. Builds can
then run on any Java version and hit remote cache primed with builds run on
different Java versions.

* Apply Foojay Toolchains Plugin

The plugin provides a repository for downloading JVMs when they're not
already installed on the machine running the build.

* Suppress MagicNumber violation

* Move compilerOptions to kotlin block