Disable local build cache on CI (#6700)
From https://docs.gradle.org/8.5/userguide/build_cache.html#sec:build_cache_configure:
> Gradle tries to load build outputs from the local build cache first, and
> then tries the remote build cache if no build outputs are found. If
> outputs are found in the remote cache, they are also stored in the local
> cache, so next time they will be found locally.

Inversely this means that that when local cache is hit, the outputs will
not be stored in the remote cache. This is not what we want on CI nodes.
Forcing remote build cache only on CI means cache will be updated for every
task that does not have outputs already stored in the remote cache. This
ensures they're available for local dev builds.