Remove getLineAndColumnInPsiFile function
This was introduced to work around issues with mutated PSI. This is no
longer an issue since #7206. Because the formatting rule set operates on a
copy of the PsiFile and not the original PsiFile the original PsiFile will
never be modified, so the protection offered by this function is not
needed.