Remove unused function (#7499)
This function was used in the begining of detekt to tests the rules. But we find out that have the code that we are testing on another file wasn't good so we moved all those test to the new `Rule.lint(String)` (and friends). Now we can remove this. If someone out there is using this it's easy for them to replicate or a good excuse to stop using that bad pattern.