Remove usage of DEVELOCITY_ACCESS_KEY env var in build scripts (#7544)
The gradle action generates short-lived access tokens that are used when
Gradle is run.

See here for more:
https://github.com/gradle/actions/blob/v4.0.0/docs/setup-gradle.md#short-lived-access-tokens

The problem is that the short-lived access token is stored in
DEVELOCITY_ACCESS_KEY env var, which is read by settings script, so any
change to the env var value invalidates the configuration cache.

Since the short-lived token is generated on every Gradle invocation, the
configuration cache will always be invalidated, harming performance on
CI on main branch.

This commit stops reading the env var value. Develocity will reject
pushes that do not have the correct secret set so this is no less
secure.