Don't set level in reportingDescriptor if not set on rule (#7682)
Error is too strong a default for a linter in the context of other tools
that output SARIF reports. The SARIF spec says "error" level is used when
the rule was evaluated and "a serious problem was found".

If a severity is set explicitly in detekt config for a ruleset or rule then
this will be set explicitly in SARIF report as well so they're aligned.