Decouple DGP compilation version from the rest of the project (#7798)
Required for Kotlin 2.1.0. Gradle's Kotlin build scripts are compiled with
Kotlin 1.9.x as of Gradle 8.3: https://docs.gradle.org/8.10.1/userguide/compatibility.html#kotlin

If DGP is compiled with Kotlin 2.1.0 the metadata version is not readable
by Kotlin 1.9.x. Kotlin is only forward compatible with +1 version of
Kotlin, see https://kotlinlang.org/docs/kotlin-evolution-principles.html#evolving-the-binary-format

> the binary format is mostly forwards compatible with the next language
> release, but not later ones (in the cases when new features are not used,
> for example, 1.9 can understand most binaries from 2.0, but not 2.1)

To mitigate this issue we switch to the Build Tools API available in newer
Kotlin versions. This allows applying one version of KGP and compiling with
a different version of Kotlin.

If this approach does not work out (it requires two opt ins!) then we can
explore other approaches, however this works for now.