Disable detekt-compiler-plugin tests
detekt-compiler-plugin must build against detekt-compiler, but the
kotlin-compile-testing helper (and its fork) require compiler plugins to be
built against kotlin-compiler-embeddable.

There is still one functional test that can be executed by running the
:detekt-compiler-plugin:testPluginKotlinc task. There is risk of future
breakage but given the compiler-plugin is experimental and will not support
semantic analysis with the Kotlin Analysis API the impact of inadvertently
breaking something that would be picked up in these tests can be managed
through issues that are raised.