Migrate from kotlin-compiler-embeddable to kotlin-compiler (#8029)
* Switch from kotlin-compiler-embeddable to kotlin-compiler

* Update IntelliJ imports to non-relocated packages

* Optimise imports

* Repackage ktlint with relocated packages

Package relocation is necessary to allow ktlint to work with the
kotlin-compiler dependency.

* Avoid unnecessary package relocation

This relocation was only required to make detekt dependencies that used
kotlin-compiler-embeddable compatible with the compiler itself.

* Use kotlin-scripting-compiler dependency for kotlin-main-kts

The scripting-compiler-embeddable dependency used by kotlin-main-kts is
not compatible with the dependencies required by the Analysis API. The non-
embeddable kotlin-scripting-compiler lib works with the current
kotlin-scripting version of test snippet compilation and will work with
Analysis API as well so both can be used in parallel in future.

* Disable detekt-compiler-plugin tests

detekt-compiler-plugin must build against detekt-compiler, but the
kotlin-compile-testing helper (and its fork) require compiler plugins to be
built against kotlin-compiler-embeddable.

There is still one functional test that can be executed by running the
:detekt-compiler-plugin:testPluginKotlinc task. There is risk of future
breakage but given the compiler-plugin is experimental and will not support
semantic analysis with the Kotlin Analysis API the impact of inadvertently
breaking something that would be picked up in these tests can be managed
through issues that are raised.