Fix packaging of Kotlin Analysis API dependencies (#8097)
* Declare Analysis API repo for all projects

This fixes resolution for local builds when projects start using the API.

* Setup dependencies directly in analysis-api project

There's no benefit to using a bundle as it's only used in one project. This
also makes it easier to separate api & implementation dependencies.

* Explicitly declare required bundling attribute for Caffeine dependency

* Set target compatibility to Java 8

* Only publish shadow variant

There is no benefit to publishing the standard variants as they are empty
and their dependencies cannot be resolved unless downstream consumers
declare the non-standard kotlin-ide-plugin-dependencies repository.

* Don't remove the "-all" archive classifier

Although this is the only published variant, the "-all" archive classifier
is the correct convention to use to declare the published artifact bundles
its dependencies.

* Declare analysis-api and analysis-api-standalone as API dependencies

This is required to use Analysis API in downstream projects.