Bump DGP API & language versions
This is the actual minimum for Gradle 7.6.3 compatibility. 7.6.3
embeds Kotlin 1.7.10. API version should match this, while the
language version can be 1.8 as Kotlin is forwards compatible +1
language version.

Also bump compiler version to 2.1.20. This is forwards compatible
with Kotlin 2.2.0 which Gradle uses in Gradle 9, while also
supporting API version 1.7. Not sure if older API versions are
dropped in Kotlin 2.2.0 or just language versions... more testing
required.