Use Kotlin language version 1.8 in DGP (#8347)
* Use Kotlin language version 1.8 in DGP

We support Gradle 7.6.3 which embeds Kotlin 1.7.10. Kotlin is
forwards compatible +1 major version, so it will be compatible with
language version 1.8.

DGP is currently built with Kotlin 2.0.21 and no explicit language
version, so it is defaulting to language version 2.0 which has no
compatibility guarantees for Kotlin 1.7.

This was not previously identified as Gradle 7.6.3 does not support
the skipMetadataVersionCheck property which was introduced in
Gradle 8.5.

* Remove usage of data object

This was introduced in language version 1.9. DGP is now built with
1.8.