Create a DAGP dependency bundle for junit-jupiter
junit-jupiter is added by Gradle's test suite plugin when applying
`useJUnitJupiter()`. That dependency contains no classes and adds
junit-jupiter-api and junit-jupiter-params as transitive dependencies
on the test compilation classpath.

Technically it's more accurate to add explicit dependencies only on
junit-jupiter-api (and junit-jupiter-params if required) but it's much
more convenient to rely on the test suite configuration and create this
bundle so Dependency Analysis Gradle Plugin stops providing advice
about this.