Switch ForbiddenClassName to use case-sensitive glob patterns
Previously, `ForbiddenClassName` entries were interpreted as
case-insensitive substrings. This behavior was surprising, inflexible,
and not in keeping with other similar rules. This change makes the rule
interpret its entries as glob patterns using the `simplePatternToRegex`
util.

This is a breaking change. I couldn't find a 2.0.0 branch, but this
should only be included in 2.0.0.

Implements #8406