Add compilation output to friendPaths by default
If files containing internal declarations are excluded from the collection of source files being analysed there will be warnings that
internal declarations cannot be accessed.

This is because if a file with an internal declaration is excluded from
analysis it will be on the classpath (because compilation outputs are
included on the classpath) but the compiler will consider it to be
coming from an external module as it's not added to the list of friend
paths. This commit updates this so that compilation outputs are also
included on the list of friend paths, so internal declarations are
accessible during analysis.