Set default jvmTarget using compilerOptions from kotlin extension (#8755)
* Set default jvmTarget using compilerOptions from kotlin extension

It's possible for the toolchain version and jvmTarget to differ. If a
toolchain is not configured Gradle still uses one but sets the Java version
to the version that is currently running Gradle, which may not be the same
as the jvmTarget.

detekt should set jvmTarget based on the value set in the extension and not
the toolchain version. When the versions differ, the jvmTarget version is
most appropriate and should be used.

* Remove unnecessary configuration of jvmTarget