Compile DGP against kotlin-plugin-api only not KGP itself (#8756)
* Compile DGP against kotlin-plugin-api only not KGP itself

This removes the compileOnly dependency on Kotlin Gradle Plugin and updates
DGP to use what's available in kotlin-plugin-api. This is not strictly
necessary now, as DGP only uses public API from KGP which should also be
maintained going forward.

However there is one exception - obtaining the explicit API mode - which
can be found in KGP but only by accessing a deprecated interface
(KotlinTopLevelExtensionConfig). KotlinBaseExtension should be used instead
but that's only available from KGP 2.1, so the minimum compatible KGP
version will be 2.1.

DGP hasn't had an explicit minimum version in the past, but this will also
make it easier to determine that going forward.

* Disable consistent resolution in DGP

Different versions of kotlin-gradle-api are used in compilation and test
runtime.