Disable test affected by change in compiler behaviour
The nullability of expression `if (fizz % 2 == 0) fizz else 1` has
changed since Kotlin 2.3.0-Beta2. In earlier Kotlin versions this was
not nullable, but in Kotlin 2.3.0 this is now nullable. I don't see a
way to work around this behaviour change.

I _think_ this was introduced by https://youtrack.jetbrains.com/issue/KT-78127