Kotlin 2.3.0 (#8743)
* Kotlin 2.3.0

* Add pluginId property to DetektCompilerPluginRegistrar

* Remove workaround for KT-71706

* chore(deps): update plugin dev.drewhamilton.poko to v0.21.0

* Use @@ delimiter in MissingSuperCall code samples

* Disable test affected by change in compiler behaviour

The nullability of expression `if (fizz % 2 == 0) fizz else 1` has
changed since Kotlin 2.3.0-Beta2. In earlier Kotlin versions this was
not nullable, but in Kotlin 2.3.0 this is now nullable. I don't see a
way to work around this behaviour change.

I _think_ this was introduced by https://youtrack.jetbrains.com/issue/KT-78127

---------

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>